Move miscelaneous helpers to client/lib/utils/
