+++ client/lib/utils/fireGlobalEvent.ts
@@ -1,6 +1,6 @@
 import { Tracker } from 'meteor/tracker';
 
-import { settings } from '../../app/settings/client';
+import { settings } from '../../../app/settings/client';
 
 export const fireGlobalEvent = (eventName: string, detail: unknown): void => {
 	window.dispatchEvent(new CustomEvent(eventName, { detail }));
