+++ client/lib/utils/getUserEmailVerified.ts
@@ -1,4 +1,4 @@
-import type { IUser } from '../../definition/IUser';
+import type { IUser } from '../../../definition/IUser';
 
 export const getUserEmailVerified = (user: IUser): boolean | undefined =>
 	Array.isArray(user.emails) ? user.emails.find(({ verified }) => !!verified)?.verified : undefined;
