+++ client/lib/utils/goToRoomById.ts
@@ -1,10 +1,10 @@
 import { memoize } from '@rocket.chat/memo';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
-import { ChatSubscription } from '../../app/models/client';
-import { call } from '../../app/ui-utils/client';
-import { roomTypes } from '../../app/utils/client';
-import { IRoom } from '../../definition/IRoom';
+import { ChatSubscription } from '../../../app/models/client';
+import { call } from '../../../app/ui-utils/client';
+import { roomTypes } from '../../../app/utils/client';
+import { IRoom } from '../../../definition/IRoom';
 
 const getRoomById = memoize((rid: IRoom['_id']) => call('getRoomById', rid));
