+++ client/lib/utils/renderMessageEmoji.ts
@@ -1,5 +1,5 @@
-import { emojiParser } from '../../app/emoji/client/emojiParser.js';
-import { IMessage } from '../../definition/IMessage';
+import { emojiParser } from '../../../app/emoji/client/emojiParser.js';
+import { IMessage } from '../../../definition/IMessage';
 
 export const renderMessageEmoji = <T extends Partial<IMessage> & { html?: string }>(
 	message: T,
