+++ client/sidebar/header/CreateChannelWithData.js
@@ -5,7 +5,7 @@ import { usePermission } from '../../contexts/AuthorizationContext';
 import { useSetting } from '../../contexts/SettingsContext';
 import { useEndpointActionExperimental } from '../../hooks/useEndpointActionExperimental';
 import { useForm } from '../../hooks/useForm';
-import { goToRoomById } from '../../lib/goToRoomById';
+import { goToRoomById } from '../../lib/utils/goToRoomById';
 import CreateChannel from './CreateChannel';
 
 const CreateChannelWithData = ({ onClose, teamId = '', reload }) => {
