+++ client/sidebar/hooks/useSidebarPaletteColor.js
@@ -2,7 +2,7 @@ import colors from '@rocket.chat/fuselage-tokens/colors';
 import { useLayoutEffect, useEffect, useMemo } from 'react';
 
 import { useSettings } from '../../contexts/SettingsContext';
-import { isIE11 } from '../../lib/isIE11';
+import { isIE11 } from '../../lib/utils/isIE11';
 
 const isInternetExplorer11 = isIE11();
