+++ client/startup/userStatusManuallySet.ts
@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor';
 
 import { callbacks } from '../../app/callbacks/client';
 import { UserStatus } from '../../definition/UserStatus';
-import { fireGlobalEvent } from '../lib/fireGlobalEvent';
+import { fireGlobalEvent } from '../lib/utils/fireGlobalEvent';
 
 /* fire user state change globally, to listen on desktop electron client */
 Meteor.startup(() => {
