+++ client/templateHelpers/renderEmoji.ts
@@ -1,5 +1,5 @@
 import { Template } from 'meteor/templating';
 
-import { renderEmoji } from '../lib/renderEmoji';
+import { renderEmoji } from '../lib/utils/renderEmoji';
 
 Template.registerHelper('renderEmoji', renderEmoji);
