+++ client/views/blocks/textParsers.js
@@ -3,7 +3,7 @@
 import { messageParser, modalParser } from '@rocket.chat/fuselage-ui-kit';
 import React from 'react';
 
-import { renderMessageBody } from '../../lib/renderMessageBody';
+import { renderMessageBody } from '../../lib/utils/renderMessageBody';
 
 // TODO: move this to fuselage-ui-kit itself
 messageParser.text = ({ text, type } = {}) => {
